+++ packages/iov-encoding/src/transactionencoder.ts
@@ -1,6 +1,6 @@
-import { isUint8Array } from "@iov/bcp";
-import { Encoding } from "@iov/encoding";
-import { JsonCompatibleValue } from "@iov/jsonrpc";
+import { Encoding } from "./encoding";
+import { JsonCompatibleValue } from "./json";
+import { isUint8Array } from "./typechecks";
 
 const prefixes = {
   string: "string:",
@@ -52,8 +52,15 @@ export class TransactionEncoder {
     ) {
       const out: any = {};
       for (const key of Object.keys(data)) {
+        const value = (data as any)[key];
+
+        // Skip dictionary entries with value `undefined`, just like native JSON:
+        // > JSON.stringify({ foo: undefined })
+        // '{}'
+        if (value === undefined) continue;
+
         // tslint:disable-next-line: no-object-mutation
-        out[key] = TransactionEncoder.toJson((data as any)[key]);
+        out[key] = TransactionEncoder.toJson(value);
       }
       return out;
     }